Testing Event Handling
I have at least two different graphs showing up - the one with the total
spending of each locality - and the other with all the spending data
from the Benchmarking & Ranking Statistics Preview Excel file. The
problem is that I can't seem to access the right page of the spreadsheet
(Benchmarks) that Bernadette asked us to look at. Also, this code still
relies on uploading the .xls file to Google Sheets and using that
conversion.